magazin/requirements.txt
djangorestframework
djangorestframework-simplejwt
django-filter
psycopg2-binary
python-decouple
Pillow
drf-yasg
django-filter
django-jazzmin
ipython